+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 * open appliaction iframe on live preview button click #26808
 
+### Fixed
+* from now, array item replacement is in one transation (before it was in two transactions: delete and insert)
+
 ## [0.6.0]
 ### Added
 * support for list fields #26780